stride-analyzer/requirements.txt
fastapi
uvicorn[standard]
Pillow
opencv-python
numpy
azure-identity
azure-ai-openai
python-multipart